move usePagination to hooks folder
